move DataClearerForegroundAppRestartPixel into its own class
